+++ routes/cars.mjs
@@ -3,6 +3,7 @@ import Car from '../models/car';
 import {
   CAN_CREATE_CAR, CAN_EDIT_CAR, CAN_GET_CAR, CAN_LIST_CAR, CAN_REMOVE_CAR,
 } from '../models/rights';
+import { csvToJson, validateCampus } from '../middlewares/csv-to-json';
 import contentNegociation from '../middlewares/content-negociation';
 import maskOutput from '../middlewares/mask-output';
 
@@ -28,6 +29,14 @@ const router = generateCRUD(Car, {
   },
   update: {
     right: CAN_EDIT_CAR,
+  },
+  batch: {
+    right: CAN_CREATE_CAR,
+    refs: ['_id', 'label'],
+    middlewares: [
+      csvToJson,
+      validateCampus,
+    ],
   },
 });