+++ tests/routes/phones.test.mjs
@@ -1,9 +1,13 @@
-import { generateAdminJWTHeader as originalGenerateAdminJWTHeader, generateDriverJWTHeader } from '../request';
+import {
+  generateAdminJWTHeader as originalGenerateAdminJWTHeader,
+  generateDriverJWTHeader,
+  generateSuperAdminJWTHeader,
+} from '../request';
 import Campus, { generateDummyCampus } from '../models/campus';
 import { createDummyPhoneModel } from '../models/phone-model';
 import Phone, { generateDummyPhone } from '../models/phone';
 import {
-  testCreate, testCreateUnicity, testDelete, testList, testGet, testUpdate,
+  testCreate, testCreateUnicity, testDelete, testList, testGet, testUpdate, testBatch,
 } from '../helpers/crud';
 
 const campus = generateDummyCampus();
@@ -75,5 +79,12 @@ describe('Test the phone API endpoint', () => {
   it(...testUpdate(Phone, {
     ...config,
     route: ({ id }) => `${config.route}/${id}`,
+  }));
+
+  it(...testBatch(Phone, {
+    ...config,
+    route: `${config.route}/batch`,
+    queryParams: {},
+    canCall: [generateSuperAdminJWTHeader],
   }));
 });